+++ druid-shell/src/platform/web/window.rs
@@ -44,8 +44,11 @@ use crate::scale::{Scale, ScaledArea};
 use crate::keyboard::{KbKey, KeyState, Modifiers};
 use crate::mouse::{Cursor, CursorDesc, MouseButton, MouseButtons, MouseEvent};
 use crate::region::Region;
+use crate::text::{simulate_input, Event};
 use crate::window;
-use crate::window::{FileDialogToken, IdleToken, TimerToken, WinHandler, WindowLevel};
+use crate::window::{
+    FileDialogToken, IdleToken, TextFieldToken, TimerToken, WinHandler, WindowLevel,
+};
 
 // This is a macro instead of a function since KeyboardEvent and MouseEvent has identical functions
 // to query modifier key states.
@@ -109,6 +112,7 @@ struct WindowState {
     context: web_sys::CanvasRenderingContext2d,
     invalid: RefCell<Region>,
     click_counter: ClickCounter,
+    active_text_input: Cell<Option<TextFieldToken>>,
 }
 
 // TODO: support custom cursors
@@ -295,7 +299,10 @@ fn setup_keydown_callback(ws: &Rc<WindowState>) {
             // Prevent the browser from going back a page by default.
             event.prevent_default();
         }
-        state.handler.borrow_mut().key_down(kb_event);
+        let mut handler = state.handler.borrow_mut();
+        if !handler.key_down(kb_event.clone()) {
+            simulate_input(&mut **handler, state.active_text_input.get(), kb_event);
+        }
     });
 }
 
@@ -436,6 +443,7 @@ impl WindowBuilder {
             context,
             invalid: RefCell::new(Region::EMPTY),
             click_counter: ClickCounter::default(),
+            active_text_input: Cell::new(None),
         });
 
         setup_web_callbacks(&window);
@@ -545,6 +553,28 @@ impl WindowHandle {
             .unwrap_or_else(|| panic!("Failed to produce a text context"));
 
         PietText::new(s.context.clone())
+    }
+
+    pub fn add_text_field(&self) -> TextFieldToken {
+        TextFieldToken::next()
+    }
+
+    pub fn remove_text_field(&self, token: TextFieldToken) {
+        if let Some(state) = self.0.upgrade() {
+            if state.active_text_input.get() == Some(token) {
+                state.active_text_input.set(None);
+            }
+        }
+    }
+
+    pub fn set_focused_text_field(&self, active_field: Option<TextFieldToken>) {
+        if let Some(state) = self.0.upgrade() {
+            state.active_text_input.set(active_field);
+        }
+    }
+
+    pub fn update_text_field(&self, _token: TextFieldToken, _update: Event) {
+        // no-op for now, until we get a properly implemented text input
     }
 
     pub fn request_timer(&self, deadline: Instant) -> TimerToken {
@@ -726,9 +756,11 @@ fn set_cursor(canvas: &web_sys::HtmlCanvasElement, cursor: &Cursor) {
         .style()
         .set_property(
             "cursor",
+            #[allow(deprecated)]
             match cursor {
                 Cursor::Arrow => "default",
                 Cursor::IBeam => "text",
+                Cursor::Pointer => "pointer",
                 Cursor::Crosshair => "crosshair",
                 Cursor::OpenHand => "grab",
                 Cursor::NotAllowed => "not-allowed",